+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 FEATURES:
 
 * **New Resource:** `gigahost_dns_record`
+* **New Resource:** `gigahost_dns_redirect`
 * **New Resource:** `gigahost_dns_zone`
 * **New Resource:** `gigahost_ssh_key`
 * **New Data Source:** `gigahost_account`